refactor(examples-next-js-app): improve setup
