+++ examples/with-next-js-app/.env.development
@@ -1,4 +1,3 @@
 NODE_ENV=development
 
-ZIMIC_SERVER_URL=http://localhost:3005
-NEXT_PUBLIC_GITHUB_API_BASE_URL=$ZIMIC_SERVER_URL/github
+GITHUB_API_BASE_URL=https://api.github.com
