+++ examples/with-next-js-app/.env.test
@@ -1,1 +1,4 @@
 NODE_ENV=test
+
+ZIMIC_SERVER_URL=http://localhost:3005
+GITHUB_API_BASE_URL=$ZIMIC_SERVER_URL/github
