+++ examples/with-next-js-app/src/app/layout.tsx
@@ -3,9 +3,6 @@ import { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import { PropsWithChildren } from 'react';
 
-import { loadInterceptors } from '../../tests/interceptors';
-import InterceptorProvider from '../providers/interceptors/InterceptorProvider';
-
 import '../styles/global.css';
 
 const inter = Inter({ subsets: ['latin'] });
@@ -15,13 +12,11 @@ export const metadata: Metadata = {
   description: 'Generated by create-next-app',
 };
 
-async function RootLayout({ children }: PropsWithChildren) {
-  await loadInterceptors();
-
+function RootLayout({ children }: PropsWithChildren) {
   return (
     <html lang="en">
       <body className={clsx(inter.className, 'bg-slate-100 flex flex-col items-center justify-center min-h-screen')}>
-        <InterceptorProvider>{children}</InterceptorProvider>
+        {children}
       </body>
     </html>
   );
