+++ examples/with-next-js-app/src/config/environment.ts
@@ -1,5 +1,5 @@
 const environment = {
-  GITHUB_API_BASE_URL: process.env.NEXT_PUBLIC_GITHUB_API_BASE_URL ?? '',
+  GITHUB_API_BASE_URL: process.env.GITHUB_API_BASE_URL ?? '',
 };
 
 export default environment;
