--- examples/with-next-js-app/src/providers/interceptors/InterceptorProvider.tsx
@@ -1,17 +1,0 @@
-'use client';
-
-import { PropsWithChildren, useEffect } from 'react';
-
-import { loadInterceptors } from '../../../tests/interceptors';
-
-type Props = PropsWithChildren;
-
-function InterceptorProvider({ children }: Props) {
-  useEffect(() => {
-    void loadInterceptors();
-  }, []);
-
-  return children;
-}
-
-export default InterceptorProvider;
